refactor: reestructurar los archivos
