+++ frontend/src/components/HeaderPreview.jsx
