+++ frontend/src/components/OptionPreview.jsx
