+++ frontend/src/components/PreviewPanel.jsx
